requirements.txt
black==25.1.0
click==8.1.8
coverage==7.7.0
Django==5.1.7
django-environ==0.12.0
django-stubs==5.1.3
djangorestframework==3.15.2
factory_boy==3.3.3
flake8==7.1.2
google-cloud-secret-manager==2.23.2
gunicorn==23.0.0
ipython==9.0.2
mypy==1.15.0
psycopg2-binary==2.9.10
pytest==8.3.5
pytest-django==4.10.0
whitenoise==6.9.0
Pillow>=10.0.0
